requirements.txt
azure-ai-documentintelligence
azure-core
openai
langchain
langchain-openai
streamlit
python-dotenv
pillow
pandas